send message adjustment

+++ pkg/db/data_model_struct.go
@@ -277,8 +277,8 @@ type LocalChatLog struct {
 	IsRead           int32  `gorm:"column:is_read" json:"isRead"`
 	Status           int32  `gorm:"column:status" json:"status"`
 	Seq              int64  `gorm:"column:seq" json:"seq"`
-	SendTime         uint32 `gorm:"column:send_time" json:"sendTime"`
-	CreateTime       uint32 `gorm:"column:create_time" json:"createTime"`
+	SendTime         int64  `gorm:"column:send_time" json:"sendTime"`
+	CreateTime       int64  `gorm:"column:create_time" json:"createTime"`
 	Ex               string `gorm:"column:ex;type:varchar(1024)" json:"ex"`
 }
 
@@ -309,9 +309,9 @@ type LocalConversation struct {
 	UnreadCount       int32  `gorm:"column:unread_count" json:"unreadCount"`
 	GroupAtType       int32  `gorm:"column:group_at_type" json:"groupAtType"`
 	LatestMsg         string `gorm:"column:latest_msg;type:varchar(1000)" json:"latestMsg"`
-	LatestMsgSendTime uint32 `gorm:"column:latest_msg_send_time" json:"latestMsgSendTime"`
+	LatestMsgSendTime int64  `gorm:"column:latest_msg_send_time" json:"latestMsgSendTime"`
 	DraftText         string `gorm:"column:draft_text" json:"draftText"`
-	DraftTextTime     uint32 `gorm:"column:draft_text_time" json:"draftTextTime"`
+	DraftTextTime     int64  `gorm:"column:draft_text_time" json:"draftTextTime"`
 	IsPinned          int32  `gorm:"column:is_pinned" json:"isPinned"`
 	Ex                string `gorm:"column:ex;type:varchar(1024)" json:"ex"`
 }
